+++ driver/lib/sessions/session.ts
@@ -1,4 +1,3 @@
-import { NATIVE_CONTEXT_NAME } from '../commands/context';
 import { FlutterDriver } from '../driver';
 import { log } from '../logger';
 
@@ -52,12 +51,6 @@ export const createSession = async function(this: FlutterDriver, sessionId, caps
               `Unsupported platformName: ${caps.platformName}`,
             );
         }
-      }
-
-      // no app starts
-      if (this.socket === null) {
-        log.info(`Starting the session as ${NATIVE_CONTEXT_NAME} context`);
-        this.setContext(NATIVE_CONTEXT_NAME);
       }
 
       return [sessionId, this.opts];